refactor test_case_delegator
