+++ dialogflow-cx/tests/test_validate_form_sample.py
@@ -20,6 +20,7 @@ from contextlib import ExitStack
 
 import mock
 import pytest
+import test_case_delegator as tcd
 from common import patch_client
 from google.api_core.operation import Operation
 from google.cloud.dialogflowcx import (
@@ -35,7 +36,6 @@ from google.cloud.dialogflowcx import (
     TestRunDifference,
     Webhook,
 )
-from sample_base import DialogflowTestCaseFailure
 from validate_form_sample import ValidateFormSample
 from webhook.main import get_webhook_uri
 
@@ -178,7 +178,7 @@ def test_basic_webhook_hermetic(differences, test_result, xfail):
         sample.initialize()
         for test_case_delegator in sample.test_case_delegators.values():
             if xfail:
-                with pytest.raises(DialogflowTestCaseFailure):
+                with pytest.raises(tcd.DialogflowTestCaseFailure):
                     test_case_delegator.run_test_case(wait=0)
             else:
                 test_case_delegator.run_test_case(wait=0)
